requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
celery>=5.5.2
sqlalchemy>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pandas>=2.0.0
pyarrow>=14.0.0
polars>=0.19.0
numpy>=1.24.0
yfinance>=0.2.33
psycopg2-binary>=2.9.9
rabbitmq-admin>=0.2
python-dateutil>=2.8.2
zstandard>=0.22.0